Fix several warnings for compiling on VS2015
* Note at the time of committing there are still some warnings in MS
  headers that you might need to suppress in a couple of files.
* 3rd party code just has the warnings suppressed for ease of merging.
* The majority of warning fixes were for local variables shadowing
  other locals, function parameters, or members. In most cases they
  weren't a problem, but in some cases it was potentially dangerous!

+++ renderdoc/common/common.cpp
@@ -132,8 +132,8 @@ bool Vec16NotEqual(void *a, void *b)
 
 bool FindDiffRange(void *a, void *b, size_t bufSize, size_t &diffStart, size_t &diffEnd)
 {
-	RDCASSERT(((unsigned long)a)%16 == 0);
-	RDCASSERT(((unsigned long)b)%16 == 0);
+	RDCASSERT(uintptr_t(a)%16 == 0);
+	RDCASSERT(uintptr_t(b)%16 == 0);
 
 	diffStart = bufSize+1;
 	diffEnd = 0;
